factor out loadAndMergeConfig()
It's useful for non-device kind of configs too.
